/* PR_START p--short-term-perf 29 */ Move static methods to `SemanticModelHelper`.
